services/gateway/requirements.txt
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
uvicorn[standard]>=0.30,<1.0
PyJWT>=2.8,<3.0
aiosqlite>=0.20,<1.0
python-multipart>=0.0.12,<1.0
bcrypt>=4.0,<5.0

